Couple more exercises on Python
Couple more exercises on Python Programming language. 
An example of the exercise is finding out whether the number that the user is entering is prime a number: if the number is prime then it will output that the number is a prime. Otherwise, it will output not a prime number.